requirements.txt
streamlit
python-dotenv
langchain
pyPDF2
chromadb
google-generativeai
